fix several bugs from poi 3.7, 3.8
git-svn-id: http://npoi.googlecode.com/svn/trunk@14 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ main/SS/UserModel/DateUtil.cs
@@ -266,7 +266,7 @@ namespace NPOI.SS.UserModel
             }
             DateTime startdate = new DateTime(startYear, 1, 1);
             startdate = startdate.AddDays(wholeDays + dayAdjust - 1);
-            int millisecondsInDay = (int)((date - wholeDays) *
+            double millisecondsInDay = (int)((date - wholeDays) *
                                           DAY_MILLISECONDS + 0.5);
             return startdate.AddMilliseconds(millisecondsInDay);
         }
